\name{beamd}
\alias{beamd}
\docType{data}
\title{Infrasonic Signal from a Nuclear Explosion 
}
\description{Infrasonic signal from a nuclear explosion.   
}
\usage{data(beamd)}
\format{
 A data frame with 2048 observations (rows) on 3 numeric variables (columns):  \code{sensor1}, \code{sensor2},  \code{sensor3}. 
}
\details{This is a data frame consisting of three columns (that are not time series objects). The data are an infrasonic signal from a nuclear explosion observed at sensors on a triangular array.
}
\references{You can find demonstrations of  astsa capabilities at  
\href{https://github.com/nickpoison/astsa/blob/master/fun_with_astsa/fun_with_astsa.md}{FUN WITH ASTSA}.

The most recent version of the package can be found  at \url{https://github.com/nickpoison/astsa/}.

In addition, the News and ChangeLog files are at  \url{https://github.com/nickpoison/astsa/blob/master/NEWS.md}.

The webpages for the texts and some help on using R for time series analysis can be found at
 \url{https://nickpoison.github.io/}.  
}
\keyword{datasets}
