\name{birth}
\alias{birth}
\docType{data}
\title{U.S. Monthly Live Births 
}
\description{Monthly live births (adjusted) in thousands for the United States, 1948-1979.
}
\format{
  The format is:
 Time-Series [1:373] from 1948 to 1979: 295 286 300 278 272 268 308 321 313 308 ...
}
\references{You can find demonstrations of  astsa capabilities at  
\href{https://github.com/nickpoison/astsa/blob/master/fun_with_astsa/fun_with_astsa.md}{FUN WITH ASTSA}.

The most recent version of the package can be found  at \url{https://github.com/nickpoison/astsa/}.

In addition, the News and ChangeLog files are at  \url{https://github.com/nickpoison/astsa/blob/master/NEWS.md}.

The webpages for the texts are \url{https://www.stat.pitt.edu/stoffer/tsa4/} and \url{https://www.stat.pitt.edu/stoffer/tsda/}.  
}
\keyword{datasets}
