\name{bnrf1hvs}
\alias{bnrf1hvs}
\docType{data}
\title{Nucleotide sequence - BNRF1 of Herpesvirus saimiri
} 
\description{Nucleotide sequence of the BNRF1 gene of the herpesvirus saimiri (HVS): 1=A, 2=C, 3=G, 4=T. The data are used in Chapter 7.
}
\format{
  The format is:
 Time-Series [1:3741] from 1 to 3741: 1 4 3 2 4 4 3 4 4 4 ...
}
\references{You can find demonstrations of  astsa capabilities at  
\href{https://github.com/nickpoison/astsa/blob/master/fun_with_astsa/fun_with_astsa.md}{FUN WITH ASTSA}.

The most recent version of the package can be found  at \url{https://github.com/nickpoison/astsa/}.

In addition, the News and ChangeLog files are at  \url{https://github.com/nickpoison/astsa/blob/master/NEWS.md}.

The webpages for the texts and some help on using R for time series analysis can be found at
 \url{https://nickpoison.github.io/}.   
}
\examples{
head(bnrf1hvs)
head(dna2vector(bnrf1hvs))
}
\keyword{datasets}
