\name{climhyd}
\alias{climhyd}
\docType{data}
\title{Lake Shasta inflow data
}
\description{Lake Shasta inflow data.  This is a data frame.
}
\format{A data frame with 454 observations (rows) on the following 6 numeric variables (columns): \code{Temp}, \code{DewPt}, \code{CldCvr}, \code{WndSpd}, \code{Precip}, \code{Inflow}.
}
\details{The data are 454 months of measured values for the climatic variables: air temperature, dew point, cloud cover, wind speed, precipitation, and inflow, at Lake Shasta, California.  The man-made lake is famous for the placard stating, "We don't swim in your toilet, so don't pee in our lake."
}
\references{You can find demonstrations of  astsa capabilities at  
\href{https://github.com/nickpoison/astsa/blob/master/fun_with_astsa/fun_with_astsa.md}{FUN WITH ASTSA}.

The most recent version of the package can be found  at \url{https://github.com/nickpoison/astsa/}.

In addition, the News and ChangeLog files are at  \url{https://github.com/nickpoison/astsa/blob/master/NEWS.md}.

The webpages for the texts and some help on using R for time series analysis can be found at
 \url{https://nickpoison.github.io/}. 
}
\keyword{datasets}
