\name{cmort}
\alias{cmort}
\docType{data}
\title{Cardiovascular Mortality from the LA Pollution study
} 
\description{Average weekly cardiovascular mortality in Los Angeles County; 508 six-day smoothed averages obtained by filtering daily values over the 10 year period 1970-1979.
}
\format{
  The format is:
 Time-Series [1:508] from 1970 to 1980: 97.8 104.6 94.4 98 95.8 ...
}
\references{You can find demonstrations of  astsa capabilities at  
\href{https://github.com/nickpoison/astsa/blob/master/fun_with_astsa/fun_with_astsa.md}{FUN WITH ASTSA}.

The most recent version of the package can be found  at \url{https://github.com/nickpoison/astsa/}.

In addition, the News and ChangeLog files are at  \url{https://github.com/nickpoison/astsa/blob/master/NEWS.md}.

The webpages for the texts and some help on using R for time series analysis can be found at
 \url{https://nickpoison.github.io/}. 
}
\seealso{\code{\link{lap}}, \code{\link{lap.xts}}}
\keyword{datasets}
