\name{cpg}
\alias{cpg}
\docType{data}
\title{Hard Drive Cost per GB 
} 
\description{Median annual cost per gigabyte (GB) of storage.
}
\format{
  The format is:
 Time-Series [1:29] from 1980 to 2008:  213000.00 295000.00 260000.00 175000.00 160000.00  ...
}
\details{The median annual cost of hard drives used in computers. The data are retail prices per GB taken from a sample of manufacturers. 

We don't remember the source of this data although it's written down somewhere (we hope). An internet search on 'median annual cost per gigabyte of storage' will pull up various sites with similar information over different time periods.
}
\references{You can find demonstrations of  astsa capabilities at  
\href{https://github.com/nickpoison/astsa/blob/master/fun_with_astsa/fun_with_astsa.md}{FUN WITH ASTSA}.

The most recent version of the package can be found  at \url{https://github.com/nickpoison/astsa/}.

In addition, the News and ChangeLog files are at  \url{https://github.com/nickpoison/astsa/blob/master/NEWS.md}.

The webpages for the texts and some help on using R for time series analysis can be found at
 \url{https://nickpoison.github.io/}. 
}
\keyword{datasets}
