\name{EQ5}
\alias{EQ5}
\docType{data}
\title{Seismic Trace of Earthquake number 5 
}
\description{Seismic trace of an earthquake [two phases or arrivals along the surface, the primary wave (\eqn{t = 1,\dots,1024}) and the shear wave (\eqn{t = 1025,\dots,2048})] recorded at a seismic station.
}
\format{
  The format is:
 Time-Series [1:2048] from 1 to 2048: 0.01749 0.01139 0.01512 0.01477 0.00651 ...
}
\references{You can find demonstrations of  astsa capabilities at  
\href{https://github.com/nickpoison/astsa/blob/master/fun_with_astsa/fun_with_astsa.md}{FUN WITH ASTSA}.

The most recent version of the package can be found  at \url{https://github.com/nickpoison/astsa/}.

In addition, the News and ChangeLog files are at  \url{https://github.com/nickpoison/astsa/blob/master/NEWS.md}.

The webpages for the texts are \url{https://www.stat.pitt.edu/stoffer/tsa4/} and \url{https://www.stat.pitt.edu/stoffer/tsda/}.  
}
\seealso{\code{\link{eqexp}}}
\keyword{datasets}
