\name{flu}
\alias{flu}
\docType{data}
\title{Monthly pneumonia and influenza deaths in the U.S., 1968 to 1978.
} 
\description{Monthly pneumonia and influenza deaths per 10,000 people in the United States for 11 years, 1968 to 1978.
}
\usage{data(flu)}
\format{
  The format is:
 Time-Series [1:132] from 1968 to 1979: 0.811 0.446 0.342 0.277 0.248 ...
}
\references{You can find demonstrations of  astsa capabilities at  
\href{https://github.com/nickpoison/astsa/blob/master/fun_with_astsa/fun_with_astsa.md}{FUN WITH ASTSA}.

The most recent version of the package can be found  at \url{https://github.com/nickpoison/astsa/}.

In addition, the News and ChangeLog files are at  \url{https://github.com/nickpoison/astsa/blob/master/NEWS.md}.

The webpages for the texts and some help on using R for time series analysis can be found at
 \url{https://nickpoison.github.io/}.  
}
\keyword{datasets}