\name{gdp}
\alias{gdp}
\docType{data}
\title{Quarterly U.S. GDP
} 
\description{Seasonally adjusted quarterly U.S. GDP from 1947(1) to 2018(3).
}
\format{
  The format is:
 Time-Series [1:287] from 1947 to 2018: 2033 2028 2023 2055 2086 ...
}
\source{\url{https://tradingeconomics.com/united-states/gdp}
}
\references{You can find demonstrations of  astsa capabilities at  
\href{https://github.com/nickpoison/astsa/blob/master/fun_with_astsa/fun_with_astsa.md}{FUN WITH ASTSA}.

The most recent version of the package can be found  at \url{https://github.com/nickpoison/astsa/}.

In addition, the News and ChangeLog files are at  \url{https://github.com/nickpoison/astsa/blob/master/NEWS.md}.

The webpages for the texts and some help on using R for time series analysis can be found at
 \url{https://nickpoison.github.io/}. 
}
\keyword{datasets}