\name{GDP23}
\alias{GDP}
\docType{data}
\title{Quarterly U.S. GDP - updated to 2023
} 
\description{Seasonally adjusted quarterly U.S. GDP from 1947(1) to 2023(1).
}
\format{
  The format is:
 Time-Series [1:305] from 1947 to 2023: 243.164 245.968 249.585 259.745 ...
}
\source{\kbd{https://fred.stlouisfed.org/series/GDP}
}
\references{You can find demonstrations of  astsa capabilities at  
\href{https://github.com/nickpoison/astsa/blob/master/fun_with_astsa/fun_with_astsa.md}{FUN WITH ASTSA}.

The most recent version of the package can be found  at \url{https://github.com/nickpoison/astsa/}.

In addition, the News and ChangeLog files are at  \url{https://github.com/nickpoison/astsa/blob/master/NEWS.md}.

The webpages for the texts and some help on using R for time series analysis can be found at
 \url{https://nickpoison.github.io/}. 
}
\seealso{\link{gdp}
}
\keyword{datasets}