\name{gnp}
\alias{gnp}
\docType{data}
\title{Quarterly U.S. GNP
} 
\description{Seasonally adjusted quarterly U.S. GNP from 1947(1) to 2002(3).
}
\format{
  The format is:
 Time-Series [1:223] from 1947 to 2002: 1489 1497 1500 1524 1547 ...
}
\source{\url{https://research.stlouisfed.org/}
}
\references{You can find demonstrations of  astsa capabilities at  
\href{https://github.com/nickpoison/astsa/blob/master/fun_with_astsa/fun_with_astsa.md}{FUN WITH ASTSA}.

The most recent version of the package can be found  at \url{https://github.com/nickpoison/astsa/}.

In addition, the News and ChangeLog files are at  \url{https://github.com/nickpoison/astsa/blob/master/NEWS.md}.

The webpages for the texts and some help on using R for time series analysis can be found at
 \url{https://nickpoison.github.io/}. 
}
\seealso{\link{GNP}
}
\keyword{datasets}
