\name{GNP23}
\alias{GNP}
\docType{data}
\title{Quarterly U.S. GNP - updated to 2023
} 
\description{Seasonally adjusted quarterly U.S. GNP from 1947(1) to 2022(4).
}
\format{
  The format is:
 Time-Series [1:305] from 1947 to 2023: 244.142 247.063 250.716 260.981 ...
}
\source{\kbd{https://fred.stlouisfed.org/series/GNP}
}
\references{You can find demonstrations of  astsa capabilities at  
\href{https://github.com/nickpoison/astsa/blob/master/fun_with_astsa/fun_with_astsa.md}{FUN WITH ASTSA}.

The most recent version of the package can be found  at \url{https://github.com/nickpoison/astsa/}.

In addition, the News and ChangeLog files are at  \url{https://github.com/nickpoison/astsa/blob/master/NEWS.md}.

The webpages for the texts and some help on using R for time series analysis can be found at
 \url{https://nickpoison.github.io/}. 
}
\seealso{\link{gnp}, \link{GDP}, \link{gdp}
}
\keyword{datasets}
