\name{gtemp.month}
\alias{gtemp.month}
\docType{data}
\title{
Monthly global average surface temperatures by year
}
\description{
Monthly global average surface temperatures by year. The temperature of the air measured 2 meters above the ground, encompassing land, sea, and in-land water surfaces.
}
\format{
  A data frame with 12 monthly observations (as rows) for the years 1975-2023 (as columns in reverse order).
}
\details{
Monthly global average surface temperatures by year. The temperature of the air measured 2 meters above the ground, encompassing land, sea, and in-land water surfaces.
}
\source{
\kbd{https://ourworldindata.org/grapher/monthly-average-surface-temperatures-by-year}
}
\references{You can find demonstrations of  astsa capabilities at
\href{https://github.com/nickpoison/astsa/blob/master/fun_with_astsa/fun_with_astsa.md}{FUN WITH ASTSA}.

The most recent version of the package can be found  at \url{https://github.com/nickpoison/astsa/}.

In addition, the News and ChangeLog files are at  \url{https://github.com/nickpoison/astsa/blob/master/NEWS.md}.

The webpages for the texts and some help on using R for time series analysis can be found at
 \url{https://nickpoison.github.io/}. 
}
\keyword{datasets}
