\name{gtemp}
\alias{gtemp}
\docType{data}
\title{Global mean land-ocean temperature deviations
} 
\description{This data file is old and is here only for compatibility.  See \code{\link{globtemp}} and \code{\link{gtemp_land}}. The original description is:  Global mean land-ocean temperature deviations (from 1951-1980 average), measured in degrees centigrade, for the years 1880-2009.
}
\format{
  The format is:
 Time-Series [1:130] from 1880 to 2009: -0.28 -0.21 -0.26 -0.27 -0.32 -0.32 -0.29 -0.36 -0.27 -0.17 ...
}
\source{\kbd{https://data.giss.nasa.gov/gistemp/graphs/}
}
\references{You can find demonstrations of  astsa capabilities at
\href{https://github.com/nickpoison/astsa/blob/master/fun_with_astsa/fun_with_astsa.md}{FUN WITH ASTSA}.

The most recent version of the package can be found  at \url{https://github.com/nickpoison/astsa/}.

In addition, the News and ChangeLog files are at  \url{https://github.com/nickpoison/astsa/blob/master/NEWS.md}.

The webpages for the texts and some help on using R for time series analysis can be found at
 \url{https://nickpoison.github.io/}. 
}
\seealso{\code{\link{gtemp_land}}, \code{\link{gtemp_ocean}}, \code{\link{globtemp}}, \code{\link{globtempl}},  \code{\link{gtemp2}}}
\keyword{datasets}