\name{gtemp_both}
\alias{gtemp_both}
\docType{data}
\title{Global mean land and open ocean temperature deviations, 1850-2023
} 
\description{Annual temperature anomalies (in degress centigrade) averaged over the Earth's land and ocean area from 1850 to 2023.  Anomalies are with respect to the 1991-2020 average.
}
\format{
  The format is:
  Time-Series [1:174] from 1850 to 2023: -0.24 -0.25 -0.27 -0.15 -0.05 -0.16 -0.29 -0.32 -0.19 -0.04 ...
}
\source{\kbd{https://www.ncei.noaa.gov/access/monitoring/climate-at-a-glance/global/time-series/}
}
\references{You can find demonstrations of  astsa capabilities at
\href{https://github.com/nickpoison/astsa/blob/master/fun_with_astsa/fun_with_astsa.md}{FUN WITH ASTSA}.

The most recent version of the package can be found  at \url{https://github.com/nickpoison/astsa/}.

In addition, the News and ChangeLog files are at  \url{https://github.com/nickpoison/astsa/blob/master/NEWS.md}.

The webpages for the texts and some help on using R for time series analysis can be found at
 \url{https://nickpoison.github.io/}. 
}
\seealso{\code{\link{gtemp_land}}, \code{\link{gtemp_ocean}}}
\keyword{datasets}