\name{gtemp_land}
\alias{gtemp_land}
\docType{data}
\title{Global mean land temperature deviations, 1850-2023
} 
\description{Annual temperature anomalies (in degress centigrade) averaged over the Earth's land area from 1850 to 2023.
 Anomalies are with respect to the 1991-2020 average.
}
\format{
  The format is:
  Time-Series [1:174] from 1850 to 2023: -0.50 -0.60 -0.50 -0.50 -0.20 -0.50 -0.80 -0.40 -0.40 -0.10 ...
}
\source{\kbd{https://www.ncei.noaa.gov/access/monitoring/climate-at-a-glance/global/time-series/}
}
\references{You can find demonstrations of  astsa capabilities at
\href{https://github.com/nickpoison/astsa/blob/master/fun_with_astsa/fun_with_astsa.md}{FUN WITH ASTSA}.

The most recent version of the package can be found  at \url{https://github.com/nickpoison/astsa/}.

In addition, the News and ChangeLog files are at  \url{https://github.com/nickpoison/astsa/blob/master/NEWS.md}.

The webpages for the texts and some help on using R for time series analysis can be found at
 \url{https://nickpoison.github.io/}. 
}
\seealso{\code{\link{gtemp_ocean}}, \code{\link{gtemp_both}}}
\keyword{datasets}