\name{gtemp_ocean}
\alias{gtemp_ocean}
\docType{data}
\title{Global mean ocean temperature deviations - updated to 2021
} 
\description{Annual sea surface temperature anomalies averaged over the part of the ocean that is free of ice at all times (open ocean) from 1880 to 2021.
}
\format{
  The format is:
  Time-Series [1:142] from 1880 to 2021: -0.05 0.01 0 -0.06 -0.15 -0.21 -0.21 -0.24 -0.05 -0.04 ...
}
\source{\kbd{https://data.giss.nasa.gov/gistemp/graphs/}
}
\references{You can find demonstrations of  astsa capabilities at
\href{https://github.com/nickpoison/astsa/blob/master/fun_with_astsa/fun_with_astsa.md}{FUN WITH ASTSA}.

The most recent version of the package can be found  at \url{https://github.com/nickpoison/astsa/}.

In addition, the News and ChangeLog files are at  \url{https://github.com/nickpoison/astsa/blob/master/NEWS.md}.

The webpages for the texts and some help on using R for time series analysis can be found at
 \url{https://nickpoison.github.io/}. 
}
\seealso{\code{\link{gtemp_land}}, \code{\link{globtemp}}, \code{\link{globtempl}},  \code{\link{gtemp2}}}
\keyword{datasets}