\name{Lynx}
\alias{Lynx}
\docType{data}
\title{Canadian Lynx
}
\description{This is one of the classic studies of predator-prey interactions, the 90-year data set is the number, in thousands, of  lynx  pelts  purchased by the Hudson's Bay Company of Canada. While this is an indirect measure of predation, the assumption is that there is a direct relationship between the number of pelts collected and the number of hare and lynx in the wild.
} 
\usage{data("Lynx")}
\format{
  The format is:
 Time-Series [1:91] from 1845 to 1935: 30.1 45.1 49.1 39.5 21.2 ...
}
\note{The data are in  units of one  thousand.  This data set pairs with \code{\link{Hare}} and is NOT the same as  \code{\link{lynx}}.
}
\source{From Odum's "Fundamentals of Ecology",  p. 191. Additional information at\cr
\kbd{http://people.whitman.edu/~hundledr/courses/M250F03/M250.html}\cr 
scroll down to: Chapter 6, Difference Equations
}
\references{You can find demonstrations of  astsa capabilities at
\href{https://github.com/nickpoison/astsa/blob/master/fun_with_astsa/fun_with_astsa.md}{FUN WITH ASTSA}.

The most recent version of the package can be found  at \url{https://github.com/nickpoison/astsa/}.

In addition, the News and ChangeLog files are at  \url{https://github.com/nickpoison/astsa/blob/master/NEWS.md}.

The webpages for the texts and some help on using R for time series analysis can be found at
 \url{https://nickpoison.github.io/}. 
}
\seealso{\code{\link{Hare}}}
\keyword{datasets}
