\name{nyse}
\alias{nyse}
\docType{data}
\title{Returns of the New York Stock Exchange
} 
\description{Returns of the New York Stock Exchange (NYSE) from February 2, 1984 to December 31, 1991.
}
\format{
  The format is:
 Time-Series [1:2000] from 1 to 2000: 0.00335 -0.01418 -0.01673 0.00229 -0.01692 ...
}
\source{S+GARCH module -  Version 1.1 Release 2: 1998
}
\note{Various packages have data sets called \code{nyse}. Consequently, it may be best to specify this data set as \code{nyse = astsa::nyse} to avoid conflicts.
}
\references{You can find demonstrations of  astsa capabilities at  
\href{https://github.com/nickpoison/astsa/blob/master/fun_with_astsa/fun_with_astsa.md}{FUN WITH ASTSA}.

The most recent version of the package can be found  at \url{https://github.com/nickpoison/astsa/}.

In addition, the News and ChangeLog files are at  \url{https://github.com/nickpoison/astsa/blob/master/NEWS.md}.

The webpages for the texts and some help on using R for time series analysis can be found at
 \url{https://nickpoison.github.io/}.  
}
\keyword{datasets}
