\name{oil}
\alias{oil}
\docType{data}
\title{Crude oil, WTI spot price FOB
} 
\description{Crude oil, WTI spot price FOB (in dollars per barrel), weekly data from 2000 to mid-2010.
}
\format{
  The format is:
 Time-Series [1:545] from 2000 to 2010: 26.2 26.1 26.3 24.9 26.3 ...
}
\details{pairs with the series \code{gas}
}
\source{Data were obtained from the URL: www.eia.doe.gov/dnav/pet/pet_pri_spt_s1_w.htm
}
\references{You can find demonstrations of  astsa capabilities at  
\href{https://github.com/nickpoison/astsa/blob/master/fun_with_astsa/fun_with_astsa.md}{FUN WITH ASTSA}.

The most recent version of the package can be found  at \url{https://github.com/nickpoison/astsa/}.

In addition, the News and ChangeLog files are at  \url{https://github.com/nickpoison/astsa/blob/master/NEWS.md}.

The webpages for the texts and some help on using R for time series analysis can be found at
 \url{https://nickpoison.github.io/}.    
}
\seealso{\code{\link{gas}}}
\keyword{datasets}
