\name{part}
\alias{part}
\docType{data}
\title{Particulate levels from the LA pollution study
} 
\description{Particulate series corresponding to \code{cmort} from the LA pollution study.
}
\format{
  The format is:
 Time-Series [1:508] from 1970 to 1980: 72.7 49.6 55.7 55.2 66 ...
}
\references{You can find demonstrations of  astsa capabilities at  
\href{https://github.com/nickpoison/astsa/blob/master/fun_with_astsa/fun_with_astsa.md}{FUN WITH ASTSA}.

The most recent version of the package can be found  at \url{https://github.com/nickpoison/astsa/}.

In addition, the News and ChangeLog files are at  \url{https://github.com/nickpoison/astsa/blob/master/NEWS.md}.

The webpages for the texts and some help on using R for time series analysis can be found at
 \url{https://nickpoison.github.io/}.   
}
\seealso{\code{\link{lap}}, \code{\link{lap.xts}}}
\keyword{datasets}
