\name{qinfl}
\alias{qinfl}
\docType{data}
\title{Quarterly Inflation
} 
\description{Quarterly inflation rate in the Consumer Price Index from 1953-Ito 1980-II, n = 110 observations.
}
\format{
  The format is:
 Time-Series [1:110] from 1953 to 1980: 1.673 3.173 0.492 -0.327 -0.333 ...
}
\details{pairs with qintr (interest rate)
}
\source{Newbold, P. and T. Bos (1985). \emph{Stochastic Parameter Regression Models}. Beverly Hills: Sage.
}
\references{You can find demonstrations of  astsa capabilities at  
\href{https://github.com/nickpoison/astsa/blob/master/fun_with_astsa/fun_with_astsa.md}{FUN WITH ASTSA}.

The most recent version of the package can be found  at \url{https://github.com/nickpoison/astsa/}.

In addition, the News and ChangeLog files are at  \url{https://github.com/nickpoison/astsa/blob/master/NEWS.md}.

The webpages for the texts and some help on using R for time series analysis can be found at
 \url{https://nickpoison.github.io/}.   
}
\seealso{\code{\link{qintr}}}
\keyword{datasets}
