\name{qintr}
\alias{qintr}
\docType{data}
\title{Quarterly Interest Rate
}
\description{Quarterly interest rate recorded for Treasury bills from 1953-Ito 1980-II, n = 110 observations. 
}
\format{
  The format is:
 Time-Series [1:110] from 1953 to 1980: 1.98 2.15 1.96 1.47 1.06 ...
}
\details{pairs with qinfl (inflation)
}
\source{Newbold, P. and T. Bos (1985). \emph{Stochastic Parameter Regression Models}. Beverly Hills: Sage.
}
\references{You can find demonstrations of  astsa capabilities at  
\href{https://github.com/nickpoison/astsa/blob/master/fun_with_astsa/fun_with_astsa.md}{FUN WITH ASTSA}.

The most recent version of the package can be found  at \url{https://github.com/nickpoison/astsa/}.

In addition, the News and ChangeLog files are at  \url{https://github.com/nickpoison/astsa/blob/master/NEWS.md}.

The webpages for the texts and some help on using R for time series analysis can be found at
 \url{https://nickpoison.github.io/}.   
}
\seealso{\code{\link{qinfl}}}
\keyword{datasets}
