\name{sales}
\alias{sales}
\docType{data}
\title{Sales
} 
\description{Sales, 150 months; taken from Box and Jenkins (1970).
}
\format{
  The format is:
 Time-Series [1:150] from 1 to 150: 200 200 199 199 199 ...
}
\details{This is also the R data set \code{BJsales}: The sales time series \code{BJsales} and leading indicator \code{BJsales.lead} each contain 150 observations. The objects are of class "ts".
}
\references{You can find demonstrations of  astsa capabilities at  
\href{https://github.com/nickpoison/astsa/blob/master/fun_with_astsa/fun_with_astsa.md}{FUN WITH ASTSA}.

The most recent version of the package can be found  at \url{https://github.com/nickpoison/astsa/}.

In addition, the News and ChangeLog files are at  \url{https://github.com/nickpoison/astsa/blob/master/NEWS.md}.

The webpages for the texts and some help on using R for time series analysis can be found at
 \url{https://nickpoison.github.io/}.   
}
\seealso{\code{\link{lead}}}
\keyword{datasets}
