\name{salmon}
\alias{salmon}
\docType{data}
\title{
Monthly export price of salmon
}
\description{
Farm Bred Norwegian Salmon, export price, US Dollars per Kilogram
}
\format{
  The format is:
 Time-Series [1:166] from September 2003 to June 2017: 2.88 3.16 2.96 3.12 3.23 3.32 3.45 3.61 3.48 3.21 ...
}
\source{\url{https://www.indexmundi.com/commodities/}
}
\references{You can find demonstrations of  astsa capabilities at  
\href{https://github.com/nickpoison/astsa/blob/master/fun_with_astsa/fun_with_astsa.md}{FUN WITH ASTSA}.

The most recent version of the package can be found  at \url{https://github.com/nickpoison/astsa/}.

In addition, the News and ChangeLog files are at  \url{https://github.com/nickpoison/astsa/blob/master/NEWS.md}.

The webpages for the texts and some help on using R for time series analysis can be found at
 \url{https://nickpoison.github.io/}.   
}
\keyword{datasets}
