\name{salt}
\alias{salt}
\docType{data}
\title{Salt Profiles
} 
\description{Salt profiles taken over a spatial grid set out on an agricultural field, 64 rows at 17-ft spacing.
}
\usage{data(salt)}
\format{
  The format is:
 Time-Series [1:64] from 1 to 64: 6 6 6 3 3 3 4 4 4 1.5 ...
}
\details{pairs with \code{saltemp}, temperature profiles on the same grid
}
\references{You can find demonstrations of  astsa capabilities at  
\href{https://github.com/nickpoison/astsa/blob/master/fun_with_astsa/fun_with_astsa.md}{FUN WITH ASTSA}.

The most recent version of the package can be found  at \url{https://github.com/nickpoison/astsa/}.

In addition, the News and ChangeLog files are at  \url{https://github.com/nickpoison/astsa/blob/master/NEWS.md}.

The webpages for the texts and some help on using R for time series analysis can be found at
 \url{https://nickpoison.github.io/}.   
}
\seealso{\code{\link{saltemp}}}
\keyword{datasets}
