\name{saltemp}
\alias{saltemp}
\docType{data}
\title{Temperature Profiles 
}
\description{Temperature profiles over a spatial grid set out on an agricultural field, 64 rows at 17-ft spacing.
}
\usage{data(saltemp)}
\format{
  The format is:
 Time-Series [1:64] from 1 to 64: 5.98 6.54 6.78 6.34 6.96 6.51 6.72 7.44 7.74 6.85 ...
}
\details{pairs with \code{salt}, salt profiles on the same grid
}
\references{You can find demonstrations of  astsa capabilities at  
\href{https://github.com/nickpoison/astsa/blob/master/fun_with_astsa/fun_with_astsa.md}{FUN WITH ASTSA}.

The most recent version of the package can be found  at \url{https://github.com/nickpoison/astsa/}.

In addition, the News and ChangeLog files are at  \url{https://github.com/nickpoison/astsa/blob/master/NEWS.md}.

The webpages for the texts and some help on using R for time series analysis can be found at
 \url{https://nickpoison.github.io/}.  
}
\seealso{\code{\link{salt}}}
\keyword{datasets}
