\name{sp500.gr}
\alias{sp500.gr}
\docType{data}
\title{Returns of the S&P 500
} 
\description{Daily growth rate of the   S&P 500 from 2001 though 2011.
}
\format{
  The format is: Time Series; Start = c(2001, 2); End = c(2011, 209); Frequency = 252 
}
\source{
Douc, Moulines, & Stoffer (2014). \emph{Nonlinear Time Series: Theory, Methods and Applications with R Examples}. CRC Press.  \url{https://doi.org/10.1201/b16331}.
}
\references{You can find demonstrations of  astsa capabilities at
\href{https://github.com/nickpoison/astsa/blob/master/fun_with_astsa/fun_with_astsa.md}{FUN WITH ASTSA}.

The most recent version of the package can be found  at \url{https://github.com/nickpoison/astsa/}.

In addition, the News and ChangeLog files are at  \url{https://github.com/nickpoison/astsa/blob/master/NEWS.md}.

The webpages for the texts and some help on using R for time series analysis can be found at
 \url{https://nickpoison.github.io/}. 
}
\keyword{datasets}
