\name{speech}
\alias{speech}
\docType{data}
\title{Speech Recording
} 
\description{A small .1 second (1000 points) sample of recorded speech for the phrase "aaa...hhh".
}
\format{
  The format is:
 Time-Series [1:1020] from 1 to 1020: 1814 1556 1442 1416 1352 ...
}
\references{You can find demonstrations of  astsa capabilities at  
\href{https://github.com/nickpoison/astsa/blob/master/fun_with_astsa/fun_with_astsa.md}{FUN WITH ASTSA}.

The most recent version of the package can be found  at \url{https://github.com/nickpoison/astsa/}.

In addition, the News and ChangeLog files are at  \url{https://github.com/nickpoison/astsa/blob/master/NEWS.md}.

The webpages for the texts and some help on using R for time series analysis can be found at
 \url{https://nickpoison.github.io/}.  
}
\keyword{datasets}
