\name{sunspotz}
\alias{sunspotz}
\docType{data}
\title{Biannual Sunspot Numbers
} 
\description{Biannual smoothed (12-month moving average) number of sunspots from June 1749 to December 1978; n = 459. The "z" on the end is to distinguish this series from the one included with R (called \code{sunspots}).
}
\format{
The format is: 
Time Series:
Start = c(1749, 1) 
End = c(1978, 1) 
Frequency = 2 
}
\references{You can find demonstrations of  astsa capabilities at  
\href{https://github.com/nickpoison/astsa/blob/master/fun_with_astsa/fun_with_astsa.md}{FUN WITH ASTSA}.

The most recent version of the package can be found  at \url{https://github.com/nickpoison/astsa/}.

In addition, the News and ChangeLog files are at  \url{https://github.com/nickpoison/astsa/blob/master/NEWS.md}.

The webpages for the texts and some help on using R for time series analysis can be found at
 \url{https://nickpoison.github.io/}.   
}
\keyword{datasets}
