\name{unemp}
\alias{unemp}
\docType{data}
\title{U.S. Unemployment
}

\description{Monthly U.S. Unemployment series (1948-1978, n = 372)
}
\usage{data(unemp)}
\format{
  The format is:
 Time-Series [1:372] from 1948 to 1979: 235 281 265 241 201 ...
}
\references{You can find demonstrations of  astsa capabilities at  
\href{https://github.com/nickpoison/astsa/blob/master/fun_with_astsa/fun_with_astsa.md}{FUN WITH ASTSA}.

The most recent version of the package can be found  at \url{https://github.com/nickpoison/astsa/}.

In addition, the News and ChangeLog files are at  \url{https://github.com/nickpoison/astsa/blob/master/NEWS.md}.

The webpages for the texts and some help on using R for time series analysis can be found at
 \url{https://nickpoison.github.io/}.   
}
\seealso{\code{\link{UnempRate}}}
\keyword{datasets}
