\name{USpop}
\alias{USpop}
\docType{data}
\title{
U.S. Population - 1900 to 2010
}
\description{
U.S. Population by official census, every ten years from 1900 to 2010.
}
\format{
  The format is:
 Time-Series [1:12] from 1900 to 2010: 76 92 106 123 132 ...
}
\details{
The census from 2020 is not included in this data set because, by many accounts, 
it was a nightmare mainly due to the COVID-19 pandemic (\kbd{https://www.nytimes.com/2022/03/10/us/census-undercounted-population.html}. To make matters worse, the Census Bureau is in the Deptartment of Commerece, and its head reports directly to the POTUS, who at the time was DJ tRump: "Historians rank Trump among worst presidents in US history ... "
}
\source{
\kbd{https://www.census.gov/}
}
\references{You can find demonstrations of  astsa capabilities at
\href{https://github.com/nickpoison/astsa/blob/master/fun_with_astsa/fun_with_astsa.md}{FUN WITH ASTSA}.

The most recent version of the package can be found  at \url{https://github.com/nickpoison/astsa/}.

In addition, the News and ChangeLog files are at  \url{https://github.com/nickpoison/astsa/blob/master/NEWS.md}.

The webpages for the texts and some help on using R for time series analysis can be found at
 \url{https://nickpoison.github.io/}. 
}
\keyword{datasets}
