\name{varve}
\alias{varve}
\docType{data}
\title{Annual Varve Series
}

\description{Sedimentary deposits from one location in Massachusetts for 634 years, beginning nearly 12,000 years ago.
}
\format{
  The format is:
 Time-Series [1:634] from 1 to 634: 26.3 27.4 42.3 58.3 20.6 ...
}
\source{
Antevs, E.  (1928). The Last Glaciation with Special Reference to the Ice Retreat in Northeastern North America.
\emph{Amer. Geog. Soc. Res. Series}, No. 17, 292p.

Shumway and  Verosub (1992). State space modeling of paleoclimatic time series. \emph{Proceedings of the 5th International Meeting on Statistical Climatology}, June, 1992, University of Toronto, 139p.
}
\references{You can find demonstrations of  astsa capabilities at  
\href{https://github.com/nickpoison/astsa/blob/master/fun_with_astsa/fun_with_astsa.md}{FUN WITH ASTSA}.

The most recent version of the package can be found  at \url{https://github.com/nickpoison/astsa/}.

In addition, the News and ChangeLog files are at  \url{https://github.com/nickpoison/astsa/blob/master/NEWS.md}.

The webpages for the texts and some help on using R for time series analysis can be found at
 \url{https://nickpoison.github.io/}.   
}
\keyword{datasets}
