\name{xA_readme}
\alias{x_readme}
\docType{data}
\title{
Scripts marked with an 'x' are scheduled to be phased out
}

\description{
Scripts marked with an 'x' are scheduled to be phased out.
}
\format{
  The format is:
 chr "Scripts marked with an 'x' are scheduled to be phased out"
}
\details{
Scripts marked with an 'x' are scheduled to be phased out.
}
\source{
Scripts marked with an 'x' are scheduled to be phased out.
}
\references{You can find demonstrations of  astsa capabilities at
\href{https://github.com/nickpoison/astsa/blob/master/fun_with_astsa/fun_with_astsa.md}{FUN WITH ASTSA}.

The most recent version of the package can be found  at \url{https://github.com/nickpoison/astsa/}.

In addition, the News and ChangeLog files are at  \url{https://github.com/nickpoison/astsa/blob/master/NEWS.md}.

The webpages for the texts and some help on using R for time series analysis can be found at
 \url{https://nickpoison.github.io/}. 
}
\author{
D.S. Stoffer
}

