\name{xglobtempl}
\alias{xglobtempl}
\docType{data}
\title{Superseded by \code{gtemp_land} - Global mean land-ocean temperature deviations. 
} 
\description{This data file is old and is scheduled to be deleted.
}
\format{
  The format is:
 Time-Series [1:136] from 1880 to 2015: -0.53 -0.51 -0.41 -0.43 -0.72 -0.56 -0.7 -0.74 -0.53 -0.25 ...
}
}
\source{\kbd{https://data.giss.nasa.gov/gistemp/graphs/}
}
\references{You can find demonstrations of  astsa capabilities at
\href{https://github.com/nickpoison/astsa/blob/master/fun_with_astsa/fun_with_astsa.md}{FUN WITH ASTSA}.

The most recent version of the package can be found  at \url{https://github.com/nickpoison/astsa/}.

In addition, the News and ChangeLog files are at  \url{https://github.com/nickpoison/astsa/blob/master/NEWS.md}.

The webpages for the texts and some help on using R for time series analysis can be found at
 \url{https://nickpoison.github.io/}. 
}
\seealso{\code{\link{gtemp_land}}, \code{\link{gtemp_ocean}}, \code{\link{gtemp_both}}}
}
\keyword{datasets}