\name{climhyd}
\alias{climhyd}
\docType{data}
\title{Lake Shasta inflow data
} 
\description{Lake Shasta inflow data.  This is a data frame.
}
\format{A data frame with 454 observations (rows) on the following 6 numeric variables (columns): \code{Temp}, \code{DewPt}, \code{CldCvr}, \code{WndSpd}, \code{Precip}, \code{Inflow}.
}
\details{The data are 454 months of measured values for the climatic variables: air temperature, dew point, cloud cover, wind speed, precipitation, and inflow, at Lake Shasta, California. 

Lake Shasta is a man-made reservoir that is situated on ancestral lands of the Shasta Indian Nation. The area around the lake was once a thriving homeland for the Shasta. The construction of the reservoir submerged many of these significant cultural sites, leading to displacement and hardship for the Shasta people. The California Gold Rush drew in outsiders in the late 1840s. For the Shasta, this was a devastating process as thousands of miners who did not respect the Shasta or their homeland, operated along various waterways. Introduction of diseases and fighting against white people, rapidly reduced the number of Shasta. For more information, see \kbd{https://www.legendsofamerica.com/shasta-indians}.
}
\references{You can find demonstrations of  astsa capabilities at  
\href{https://github.com/nickpoison/astsa/blob/master/fun_with_astsa/fun_with_astsa.md}{FUN WITH ASTSA}.

The most recent version of the package can be found  at \url{https://github.com/nickpoison/astsa/}.

In addition, the News and ChangeLog files are at  \url{https://github.com/nickpoison/astsa/blob/master/NEWS.md}.

The webpages for the texts and some help on using R for time series analysis can be found at
 \url{https://nickpoison.github.io/}. 
}
\keyword{datasets}
