\name{gtemp_land}
\alias{gtemp_land}
\docType{data}
\title{Global mean land  temperature deviations - updated to 2021
} 
\description{Annual temperature anomalies (in degress centigrade) averaged over the Earth's land area from 1880 to  2021.
}
\format{
  The format is:
  Time-Series [1:142] from 1880 to 2021: -0.6 -0.39 -0.49 -0.58 -0.75 -0.76 -0.6 -0.69 -0.54 -0.26 ...
}
\source{\kbd{https://data.giss.nasa.gov/gistemp/graphs/}
}
\references{You can find demonstrations of  astsa capabilities at
\href{https://github.com/nickpoison/astsa/blob/master/fun_with_astsa/fun_with_astsa.md}{FUN WITH ASTSA}.

The most recent version of the package can be found  at \url{https://github.com/nickpoison/astsa/}.

In addition, the News and ChangeLog files are at  \url{https://github.com/nickpoison/astsa/blob/master/NEWS.md}.

The webpages for the texts and some help on using R for time series analysis can be found at
 \url{https://nickpoison.github.io/}. 
}
\seealso{\code{\link{gtemp_ocean}},   \code{\link{globtemp}}, \code{\link{globtempl}},  \code{\link{gtemp2}}}
\keyword{datasets}