\name{gtemp_land}
\alias{gtemp_land}
\docType{data}
\title{Global mean land  temperature deviations - updated to 2017
} 
\description{Annual temperature anomalies (in degress centigrade) averaged over the Earth's land area from 1880 to  2017.
}
\format{
  The format is:
 Time-Series [1:138] from 1880 to 2017: -0.62 -0.45 -0.47 -0.62 -0.82 ...
}
\source{\kbd{https://data.giss.nasa.gov/gistemp/graphs/}
}
\references{You can find demonstrations of  astsa capabilities at
\href{https://github.com/nickpoison/astsa/blob/master/fun_with_astsa/fun_with_astsa.md}{FUN WITH ASTSA}.

The most recent version of the package can be found  at \url{https://github.com/nickpoison/astsa/}.

In addition, the News and ChangeLog files are at  \url{https://github.com/nickpoison/astsa/blob/master/NEWS.md}.

The webpages for the texts and some help on using R for time series analysis can be found at
 \url{https://nickpoison.github.io/}. 
}
\seealso{\code{\link{gtemp_ocean}},   \code{\link{globtemp}}, \code{\link{globtempl}},  \code{\link{gtemp2}}}
\keyword{datasets}