\name{gtemp_ocean}
\alias{gtemp_ocean}
\docType{data}
\title{Global mean ocean temperature deviations, 1850-2023
} 
\description{Annual sea surface temperature anomalies averaged over the part of the ocean that is free of ice at all times (open ocean) from 1850 to 2023.  Anomalies are with respect to the 1991-2020 average.
}
\format{
  The format is:
  Time-Series [1:174] from 1850 to 2023: -0.12 -0.08 -0.14  0.04  0.04  0.00 -0.05 -0.27 -0.09  0.01 ...
}
\source{\kbd{https://www.ncei.noaa.gov/access/monitoring/climate-at-a-glance/global/time-series/}
}
\references{You can find demonstrations of  astsa capabilities at
\href{https://github.com/nickpoison/astsa/blob/master/fun_with_astsa/fun_with_astsa.md}{FUN WITH ASTSA}.

The most recent version of the package can be found  at \url{https://github.com/nickpoison/astsa/}.

In addition, the News and ChangeLog files are at  \url{https://github.com/nickpoison/astsa/blob/master/NEWS.md}.

The webpages for the texts and some help on using R for time series analysis can be found at
 \url{https://nickpoison.github.io/}. 
}
\seealso{\code{\link{gtemp_land}}, \code{\link{gtemp_both}}}
\keyword{datasets}